+++ allocation.hpp
@@ -1,6 +1,6 @@
 
-int** alloc_2d_init(int r, int c);
+float** alloc_2d_init(int r, int c);
 
-void data_init(int** A, int r, int c);
+void data_init(float** A, int r, int c);
 
-void print_helper(int** A, int r, int c);
+void print_helper(float** A, int r, int c);
